Alterations

Description

Module for the realisation of Alterations of LI1 in 2023/24.

Documentation

data Tecpress

Constructors

Tecpress

Fields

pressao :: Bool

acao :: Maybe Acao

presse :: Bool

pressd :: Bool

hsalt :: Bool

type Paridade = (Bool, Bool)

type MinSeg = (Int, Int)

alterations :: a -> a

This document presents specific functions for the proper functioning of the graphics part of the game, including new functions for the definition in Gloss and changes to functions already defined in previous tasks, but with specificity for some cases of the graphics part. for some cases in the graphics area.

aoPos: Efficient function for obtaining a player's position by providing a player.

redondo: Function that rounds the values of a position to the integer value

virgulado: Function that rounds the values of a position in integers to a position in Doubles

otnEscala: Function that goes from coordinates within the matrix to the actual map

ntoEscala: Function that goes from the coordinates of the top left corner of the map to the coordinates of the centre of the matrix

estrelinhas: Function that calculates the player's rating in a stage, based on the time it took them to complete it

segtoMin: Function that converts the time from seconds to minutes and seconds

parI: Function that checks whether the rounded down value of the time is even or odd

dist: Function that calculates the distance between two positions

pmP: Function that receives a position and a list of positions, and returns the position from the list closest to the singular position entered

findBlocPC: Function that receives a list of block lists and returns the centre coordinates in the Platform block matrix

findBlocPV: Function that receives a list of block lists and returns the centre coordinates in the Empty block matrix

findBlocPE: Function that receives a list of block lists and returns the centre coordinates in the Ladder block matrix

findBlocPA: Function that receives a list of block lists and returns the centre coordinates in the Trapdoor block matrix.

findBlocBC: Function that receives a list of blocks and returns the centre coordinates in the Barrier block matrix

collat: Function that checks whether the player can move left or right, and prevents movement if it is not valid

esFix: Function that helps the player get off the ladder, minimising a small bug that occurred in normal situations.

esFixHelper: Function that helps the player get off the ladder, minimising a small bug when this occurs in normal situations.

movCon: Function responsible for allowing a continuous click on a key to move the player continuously

jumpNAssist: Function responsible for preventing the player from jumping without support

vidaJog: Function that checks if the player has less than 3 lives left

remEspada: Function that removes the damage applied to the player after 10 seconds

collatI: Function that applies the colLat function to the list of enemies

subEscadaI: Function that checks if a character is on a ladder block

esFixI: Function that applies the esFix' function to the list of enemies

fallVoidIH: Function that applies the gravity effect to an enemy

esFix': Function that helps the enemy get off the ladder, minimising a small bug when this occurred in normal situations

newMov: Function that calculates the enemy's next move

colLatB: Function that checks whether an enemy can move left or right

fallVoidI: Function that applies the fallVoidIH function to the entire list of enemies

Properties:

Function "oPos":

oPos :: Personagem -> Posicao

oPos (Personagem i1 i2 (x, y) i3 i4 i5 i6 i7 i8 i9) = (x, y)

Function "redondo":

redondo :: Posicao -> (Int,Int)

redondo (x,y) = (round x, round y)

Function "virgulado":

virgulado :: (Int,Int) -> Posicao

virgulado (x,y) = (fromIntegral x, fromIntegral y)

Function "otnEscala":

otnEscala :: Posicao -> Posicao

otnEscala (x, y) = (32 * x - 400, 496 - 32 * y)

Function "ntoEscala":

ntoEscala :: Posicao -> Posicao

ntoEscala (x, y) = ((x + 400) / 32, ((-y + 496) / 32))

Function "estrelinhas":

estrelinhas :: MinSeg -> Int

estrelinhas (min,seg) = if min == 0 && 0 <= seg && seg < 60

then 3

else if 1 <= min && min <= 2 && 0 <= seg && seg < 60

then 2

else 1

Function "segToMin":

segToMin :: Tempo -> MinSeg

segToMin t = (round t `div` 60 , round t `mod` 60)

Function "parI":

parI :: Tempo -> Paridade

parI t = if odd (ceiling t) then (True,False) else (False,True)

Function "dist":

dist :: Posicao -> Posicao -> Double

dist (x1, y1) (x2, y2) = sqrt ((x2 - x1)^2 + (y2 - y1)^2)

Function "pmP":

pmP :: Posicao -> [Posicao] -> Posicao

pmP _ [] = error "Lista vazia"

pmP p0 p = foldl1' (\p1 p2 -> if dist p0 p1 < dist p0 p2 then p1 else p2) p

Function "findBlocPC":

findBlocPC :: [[Bloco]] -> [Posicao]

findBlocPC [] = []

findBlocPC m =

concatMap (\(l, bs) -> [(fromIntegral c + 0.5, fromIntegral l + 0.5) | (b, c) <- zip bs [0..], b == Plataforma]) $ zip [0..] m

Function "findBlocPV":

findBlocPV :: [[Bloco]] -> [Posicao]

findBlocPV [] = []

findBlocPV m =

concatMap (\(l, bs) -> [(fromIntegral c + 0.5, fromIntegral l + 0.5) | (b, c) <- zip bs [0..], b == Vazio]) $ zip [0..] m

Function "findBlocPE":

findBlocPE :: [[Bloco]] -> [Posicao]

findBlocPE [] = []

findBlocPE m =

concatMap (\(l, bs) -> [(fromIntegral c + 0.5, fromIntegral l + 0.5) | (b, c) <- zip bs [0..], b == Escada]) $ zip [0..] m

Function "findBlocPA":

findBlocPA :: [[Bloco]] -> [Posicao]

findBlocPA [] = []

findBlocPA m =

concatMap (\(l, bs) -> [(fromIntegral c + 0.5, fromIntegral l + 0.5) | (b, c) <- zip bs [0..], b == Alcapao]) $ zip [0..] m

Function "findBlocBC":

findBlocBC :: [[Bloco]] -> [Posicao]

findBlocBC [] = []

findBlocBC m =

concatMap (\(l, bs) -> [(fromIntegral c + 0.5, fromIntegral l + 0.5) | (b, c) <- zip bs [0..], b == Barreira]) $ zip [0..] m

Function "colLat":

colLat :: Mapa -> Personagem -> Personagem

colLat (Mapa _ _ m3) j@(Personagem (vx, vy) _ (x, y) _ _ _ _ _ _ _) =

if colD

then j { velocidade = (min 0 vx, vy) }

else if colE

then j { velocidade = (max 0 vx, vy) }

else j

where

colE = any (==lJ) (concatMap (\(x, y) -> [(i, j) | i <- [x], j <- [y .. y + 31]]) (map (\(x,y) -> (x+32,y-16)) (map otnEscala  lP)))

colD = any (==(x,y)) (concatMap (\(x, y) -> [(i, j) | i <- [x], j <- [y .. y + 31]]) (map (\(x,y) -> (x-32,y-16)) (map otnEscala  lP)))

lP = findBlocPC m3

lJ = (x,y)

Function "esFix":

esFix :: [[Bloco]] -> Personagem -> Personagem

esFix m p@(Personagem (vx,vy) i2 i3 i4 i5 i6 i7 i8 i9 i10)

| any (`elem` [i3]) l = (Personagem (vx,vy + 32) i2 i3 i4 i5 i6 i7 i8 i9 i10)

| otherwise = p

where

l = concatMap (\(x, y) -> [(i, j) | i <- [x .. x + 31], j <- [y]]) (map (\(x,y) -> (x-16,y)) (esFixHelper m))

Function "esFixHelper":

esFixHelper :: [[Bloco]] -> [Posicao]

esFixHelper m =

let l = findBlocPC m

s = map (\(x, y) -> (x, y - 1)) (findBlocPE m)

in map otnEscala (nub [(x, y) | (x, y) <- l, (x, y) `elem` s])

Function "movCon":

movCon :: Jogo -> Tecpress -> Jogo

movCon j@(Jogo m i c jog@(Personagem (vx, vy) _ _ _ _ _ _ _ _ _)) tecpress

| pressao tecpress && (jumpNAssist m jog || (acao tecpress == Just Subir) || (acao tecpress == Just Descer)) =

atualiza' [] (acao tecpress) j

| acao tecpress == Just Saltar && not (presse tecpress || pressd tecpress) =

atualiza' [] (acao tecpress) j

| acao tecpress == Just Saltar && pressd tecpress =

atualiza' [] (Just Saltar) (j { jogador = jog { velocidade = (vx+8, vy) } })

| acao tecpress == Just Saltar && presse tecpress =

atualiza' [] (Just Saltar) (j { jogador = jog { velocidade = (vx-8, vy) } })

| acao tecpress == Just Parar =

atualiza' [] (acao tecpress) j

| otherwise = j

Function "jumpNAssist":

jumpNAssist :: Mapa -> Personagem  -> Bool

jumpNAssist (Mapa _ _ b) p@(Personagem (vx, vy) _ (x,y) _ _ _ _ _ _ _)

| any (==p') b' = True

| otherwise = False

where

p' = virgulado $ redondo (oPos p)

b' = concatMap (\(x, y) -> [(i, j) | i <- [x .. x + 45], j <- [y]]) (map (\(x,y) -> (x-24,y+32)) (map otnEscala (findBlocPC b ++ findBlocPA b)))

Function "vidaJog":

vidaJog :: Jogo -> Bool

vidaJog (Jogo m i r (Personagem i1 i2 i3 i4 i5 i6 i7 v i9 i10)) = v < 3

Function "remEspada":

remEspada :: Personagem -> Personagem

remEspada p@(Personagem i1 i2 i3 i4 i5 i6 i7 i8 i9 (f,n))

| n > 0 = (Personagem i1 i2 i3 i4 i5 i6 i7 i8 i9 (f,n - 1/30))

| otherwise = (Personagem i1 i2 i3 i4 i5 i6 i7 i8 i9 (False, 0))

Function "colLatI":

colLatI :: Mapa -> [Personagem] -> [Personagem]

colLatI m i = map (colLat m) i

Function "subEscadaI":

subEscadaI :: [[Bloco]] -> Personagem -> Bool

subEscadaI b p@(Personagem i1 i2 i3 i4 i5 i6 i7 i8 i9 i10)

| any (`elem` [i3]) b' = True

| otherwise = False

where

b' = concatMap (\(x, y) -> [(i, j) | i <- [x], j <- [y .. y + 32]]) (map (\(x,y) -> (x,y-32)) (map otnEscala (findBlocPE b)))

Function "esFixI":

esFixI :: [[Bloco]] -> [Personagem] -> [Personagem]

esFixI m i = map (esFix' m) i

Function "fallVoidIH":

fallVoidIH :: Mapa -> Personagem -> Personagem

fallVoidIH (Mapa _ _ b) p@(Personagem (vx, vy) i2 _ _ _ i6 _ _ _ _)

| subEscadaI b p = p

| not (any (==p') b') && i2 == Fantasma = p { velocidade = (vx, vy - 8) }

| otherwise = p

where

p' = virgulado $ redondo (oPos p)

b' = concatMap (\(x, y) -> [(i, j) | i <- [x .. x + 45], j <- [y]]) (map (\(x,y) -> (x-24,y+32)) (map otnEscala (findBlocPC b ++ findBlocPA b)))

Function "esFix":

esFix' :: [[Bloco]] -> Personagem -> Personagem

esFix' m p@(Personagem (vx,vy) i2 i3@(x,y) i4 i5 i6 i7 i8 i9 i10)

| any (`elem` [i3]) l = (Personagem (vx,vy) i2 (x,y+64) i4 i5 i6 i7 i8 i9 i10)

| otherwise = p

where

l = concatMap (\(x, y) -> [(i, j) | i <- [x .. x + 31], j <- [y]]) (map (\(x,y) -> (x-16,y-32)) (esFixHelper m))

Function "newMov":

newMov :: Mapa -> Personagem -> Personagem

newMov m@(Mapa m1 m2 m3) i@(Personagem (v1,v2) i2 i3 i4 i5 i6 i7 i8 i9 i10)

rop>      | subEscadaI m3 i = Personagem (v1,v2+4) i2 i3 i4 i5 i6 i7 i8 i9 i10

| colLatB m i == (True,False) && i7 = Personagem (v1-4,v2) i2 i3 Oeste i5 i6 i7 i8 i9 i10

| colLatB m i == (True,True)  && i7 = Personagem (v1+4,v2) i2 i3 Este i5 i6 i7 i8 i9 i10

| otherwise = if i4 == Este then Personagem (v1+4,v2) i2 i3 Este i5 i6 i7 i8 i9 i10 else Personagem (v1-4,v2) i2 i3 Oeste i5 i6 i7 i8 i9 i10

Function "colLatB":

colLatB :: Mapa -> Personagem -> (Bool,Bool)

colLatB (Mapa _ _ m3) j@(Personagem (vx, vy) _ (x, y) _ _ _ _ _ _ _)

| colD = (True,False)

| colE = (True,True)

| otherwise = (False, False)

where

colE = any (==lJ) (concatMap (\(x, y) -> [(i, j) | i <- [x], j <- [y .. y + 31]]) (map (\(x,y) -> (x+32,y-16)) (map otnEscala (lP ++ lB))))

colD = any (==(x,y)) (concatMap (\(x, y) -> [(i, j) | i <- [x], j <- [y .. y + 31]]) (map (\(x,y) -> (x-32,y-16)) (map otnEscala (lP ++ lB))))

lP = findBlocPC m3

lB = findBlocBC m3

lJ = (x,y)

Function "fallVoidI":

fallVoidI :: Mapa -> [Personagem] -> [Personagem] -- a

fallVoidI m i = map (fallVoidIH m) i

oPos

Arguments

:: Personagem
Player character

-> Posicao
Player position

Efficient function for obtaining a player's position by supplying a player

redondo

Arguments

:: Posicao
Position in Double

-> (Int, Int)
Rounded position in Int

Function that rounds the values of a position to the integer value

virgulado

Arguments

:: (Int, Int)
Rounded position in Int

-> Posicao
Position in Double

Function that rounds values from a position in int to a position in Doubles

otnEscala

Arguments

:: Posicao
Position in the matrix

-> Posicao
Position on the map

Function that passes from coordinates within the matrix to the actual map

ntoEscala

Arguments

:: Posicao
Position in the matrix

-> Posicao
Position on the map

Function that goes from the coordinates of the top left corner of the map to the coordinates of the centre of the matrix

estrelinhas

Arguments

:: MinSeg
Time the player spent in the phase

-> Int
Player rating from 1 to 3

Function that calculates the player's rating in a stage, based on the time it took to complete it

segToMin

Arguments

:: Tempo
Time in seconds

-> MinSeg
Time in minutes and seconds

Function that converts time from seconds to minutes and seconds

parI

Arguments

:: Tempo
Time in seconds

-> Paridade
Truth value in the form of a different type of data

Function that checks whether the rounded value under the time is even or odd

dist

Arguments

:: Posicao
First position

-> Posicao
Second position

-> Double
Distance between the two positions

Function that calculates the distance between two positions

pmP

Arguments

:: Posicao
Starting position

-> [Posicao]
List of positions

-> Posicao
List position closest to starting position

Function that receives a position and a list of positions, and returns the list position closest to the singular position entered

findBlocPC

Arguments

:: [[Bloco]]
Map blocks

-> [Posicao]
List of Platform block positions

Function that receives a list of block lists and returns the centre coordinates in the Platform block matrix

findBlocPV

Arguments

:: [[Bloco]]
Map blocks

-> [Posicao]
List of Empty block positions

Function that receives a list of block lists and returns the centre coordinates in the Empty block matrix

findBlocPE

Arguments

:: [[Bloco]]
Map blocks

-> [Posicao]
List of Ladder block positions

Function that receives a list of block lists and returns the centre coordinates in the Ladder block matrix

findBlocPA

Arguments

:: [[Bloco]]
Map blocks

-> [Posicao]
List of Trapdoor block positions

Function that receives a list of block lists and returns the centre coordinates in the Trapdoor block matrix

findBlocBC

Arguments

:: [[Bloco]]
Map blocks

-> [Posicao]
List of Barrier block positions

Function that receives a list of block lists and returns the centre coordinates in the Barrier block matrix

colLat

Arguments

:: Mapa
Game map

-> Personagem
Player character

-> Personagem
Updated player character

Function that checks whether the player can move left or right, and prevents movement if it is not valid

esFixHelper

Arguments

:: [[Bloco]]
List of map blocks

-> [Posicao]
List of platform block positions on top of the last vertical ladder

Function that helps the player get off the ladder, minimising a small bug when this occurred in normal situations

esFix

Arguments

:: [[Bloco]]
List of map blocks

-> Personagem
Player character

-> Personagem
Player character after correction

movCon

Arguments

:: Jogo
Game in question

-> Tecpress
Pressing certain buttons

-> Jogo
Updated game

Function responsible for allowing a continuous click on a key to move the player continuously

jumpNAssist

Arguments

:: Mapa
Game map

-> Personagem
Player character

-> Bool
Truth value for the jump

Function responsible for preventing the player from jumping without support

vidaJog

Arguments

:: Jogo
Game in question

-> Bool
Truth value for the determined

Function that checks if the player has less than 3 lives

remEspada

Arguments

:: Personagem
Player character

-> Personagem
Updated player character

Function that removes damage from the player after 10 seconds

colLatI

Arguments

:: Mapa
Game map

-> [Personagem]
Enemies list

-> [Personagem]
Updated enemies list

Function that applies the colLat function to the list of enemies

subEscadaI

Arguments

:: [[Bloco]]
Map defenition matriz

-> Personagem
Character

-> Bool
Truth value of the character being in the position of a Ladder block

Function that checks if a character is on a ladder block

esFixI

Arguments

:: [[Bloco]]
List of map block lists

-> [Personagem]
Enemies list

-> [Personagem]
Updated enemies list

Function that applies the esFix' function to the enemies list

fallVoidIH

Arguments

:: Mapa
Game map

-> Personagem
Enemy in question

-> Personagem
Updated enemy

Function that applies the effect of gravity to an enemy

esFix'

Arguments

:: [[Bloco]]
List of map block lists

-> Personagem
Enemy in question

-> Personagem
Updated enemy

Function that helps the enemy get off the ladder, minimising a small bug when this occurred in normal situations

newMov

Arguments

:: Mapa
Game map

-> Personagem
Enemy in question

-> Personagem
Updated enemy

Function that calculates the enemy's next move

colLatB

Arguments

:: Mapa
Game map

-> Personagem
Enemy in question

-> (Bool, Bool)
Truth value for the determined

Function that checks whether an enemy can move left or right

fallVoidI

Arguments

:: Mapa
Game map

-> [Personagem]
Enemies list

-> [Personagem]
Updated enemies list

Function that applies the fallVoidIH function to the entire list of enemies

valida' :: Jogo -> Bool

The remaining functions are changes to some of the functions defined in the previous tasks. These have been altered for specific game purposes, so we haven't done any haddock documentation on them

choquePs' :: Mapa -> [Personagem] -> Bool

alcLargo' :: Personagem -> Bool

colEntidades' :: [Personagem] -> Personagem -> [(Colecionavel, Posicao)] -> Mapa -> Bool

colEntidadesAux' :: [Personagem] -> Personagem -> [(Colecionavel, Posicao)] -> [Posicao]

movimenta' :: Semente -> Tempo -> Jogo -> Jogo

movJogador' :: Personagem -> Personagem

movInimigo' :: [Personagem] -> [Personagem]

menosAl' :: Mapa -> Personagem -> [[Bloco]]

fallVoid' :: Mapa -> Personagem -> Personagem

extEscada' :: Mapa -> [Posicao] -> [(Bloco, Bloco)]

recColec' :: Jogo -> Jogo

insColec' :: (Colecionavel, Posicao) -> Jogo -> Jogo

recColecE :: Jogo -> Bool

interColec :: Posicao -> Posicao -> Bool

randInimigos' :: Mapa -> Semente -> [Personagem] -> [Personagem]

subEscadaB :: [[Bloco]] -> Personagem -> Bool

perderVida' :: [Personagem] -> Personagem -> [Personagem]

hitDano' :: Personagem -> [Posicao]

danoDuro' :: [Personagem] -> Personagem -> Personagem

colisoesDano' :: [Personagem] -> Personagem -> [Bool]

colisoesPersonagens' :: Personagem -> Personagem -> Bool

atualiza' :: [Maybe Acao] -> Maybe Acao -> Jogo -> Jogo

atualizaPer' :: Mapa -> Maybe Acao -> Personagem -> Personagem

atualizaAcao' :: Mapa -> Acao -> Personagem -> Personagem

subEscada' :: [[Bloco]] -> Personagem -> Personagem

subVazio :: [[Bloco]] -> Personagem -> Bool

parOp' :: [[Bloco]] -> Personagem -> Bool