Task1

Description

Module for the realisation of Task1 of LI1 in 2023/24.

Documentation

colisoesParede

Arguments

:: Mapa
Game Map

-> Personagem
Details of the character in questiona

-> Bool
Logical value of the collision between map boundaries or platforms and the character

The colisoesParede function must test whether a character is colliding with any of the map's boundaries (sides or top) or with any platform blocks. To obtain this result, we use auxiliary functions that perform the various operations required, such as:

acC: function that centres the positions of the game entities

acP: function that centres the game entities

acH: function that centres the hitboxes of game entities

colPar: function that tests whether a character collides with the map boundaries

colBloco: function that tests whether a character collides with a platform block

hitPersona: function that calculates a character's hitbox, defined by a pair of positions, these being the bottom left and top right, respectively

hitBlocos: function that calculates the hitbox of a list of platforms

findBlocP: function that receives the map definition matrix and returns a list with the coordinates of the top left corner of each plataform block

Properties:

Main function "colisoesParede":

colisoesParede :: Mapa -> Personagem -> Bool

colisoesParede (Mapa _ _ m) p

| colPar m (acP p) || colBloco (hitBlocos $ findBlocP m) (hitPersona (acP p)) = True

| otherwise = False

Auxiliary function "acC":

acC :: Posicao -> Posicao

acC (x, y) = (ar x, ar y)

where

ar :: Double -> Double

ar n

| dP == 0.5 = n

| dP >= 0.9 = fromIntegral (ceiling n) - 0.5

| otherwise = fromIntegral (floor n) + 0.5

where

dP = n - fromIntegral (floor n)

Auxiliary function "acP":

acP :: Personagem -> Personagem

acP (Personagem i1 i2 i3 i4 i5 i6 i7 i8 i9 i10) = Personagem i1 i2 (acC i3) i4 i5 i6 i7 i8 i9 i10

Auxiliary function "acH":

acH :: Hitbox -> Hitbox

acH ((a,b),(c,d)) = (acC (a,b), acC (c,d))

Auxiliary function "colPar":

colPar :: [[Bloco]] -> Personagem -> Bool

colPar m (Personagem _ _ (x, y) _ (g, c) _ _ _ _ _)

| x - (g / 2) <= 0 || x + (g / 2) >= th || y - (c / 2) <= 0 || y + (c / 2) >= tv = True

| otherwise = False

where

th = fromIntegral $ length $ head m

tv = fromIntegral $ length m

Auxiliary function "colBloco":

colBloco :: [Hitbox] -> Hitbox -> Bool

colBloco [] _ = False

colBloco (((xi1, yi1), (xi2, yi2)):xs) ((xe1, yd1), (xe2, yd2)) =

elem True [x >= min xe1 xe2 && x <= max xe1 xe2 && y >= min yd1 yd2 && y <= max yd1 yd2 | (x, y) <- v] || colBloco xs ((xe1, yd1), (xe2, yd2))

where

v = [(x, y) | x <- [xi1, xi2], y <- [yi1, yi2]]

Auxiliary function "findBlocP":

findBlocP :: [[Bloco]] -> [Posicao]

findBlocP [] = []

findBlocP m =

concatMap (\(l, bs) -> [(fromIntegral c, fromIntegral l) | (b, c) <- zip bs [0..], b == Plataforma]) $ zip [0..] m

Auxiliary function "hitBlocos":

hitBlocos :: [Posicao] -> [Hitbox]

hitBlocos = map (\(x, y) -> ((x, y + 1), (x + 1, y)))

Auxiliary function "hitPersona":

hitPersona :: Personagem -> Hitbox

hitPersona (Personagem _ _ (x,y) _ (g,c) _ _ _ _ _)

= ((xe1,yd1),(xe2,yd2))

where

xe1 = x - (g / 2)

yd1 = y + (c / 2)

xe2 = x + (g / 2)

yd2 = y - (c / 2)

Examples of use:

>>> colisoesParede (Mapa ((0, 0), Este) (5, 5) [[Plataforma, Plataforma, Plataforma], [Plataforma, Vazio, Plataforma], [Plataforma, Plataforma, Plataforma]]) (Personagem (1, 0) Jogador (1, 1) Este (1, 1) False False 3 0 (False, 0))
True

>>> colisoesParede (Mapa ((0, 0), Este) (5, 5) [[Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio]]) (Personagem (1, 0) Jogador (1, 1) Este (1, 1) False False 3 0 (False, 0))
False

acC

Arguments

:: Posicao
Entity position

-> Posicao
Centred position of the entity

Function that centres the positions of the game entities

acP

Arguments

:: Personagem
Entity

-> Personagem
Centred position of the entity

Function that centres the game entities

acH

Arguments

:: Hitbox
Entity hitbox

-> Hitbox
Centred entity hitbox

Function that centres the hitboxes of game entities

colPar

Arguments

:: [[Bloco]]
Matrix that defines the map

-> Personagem
Character in question

-> Bool
Logical value of the character's collision with the map boundaries

Function that tests whether a character collides with the map boundaries

colBloco

Arguments

:: [Hitbox]
List of hitboxes for each platform block

-> Hitbox
Character hitbox

-> Bool
Logical value of the platform blocks colliding with the character

Function that tests whether a character collides with a platform block

findBlocP

Arguments

:: [[Bloco]]
Map definition matrix

-> [Posicao]
List of coordinates in the top left corner of the block

Function that receives the map definition matrix and returns a list with the coordinates of the top left corner of each plataform block

hitBlocos

Arguments

:: [Posicao]
List of platform positions

-> [Hitbox]
List of platform block hitboxes

Function that calculates the hitbox of a list of platforms

hitPersona

Arguments

:: Personagem
Character in question

-> Hitbox
Hitbox of the character in question

Function that calculates a character's hitbox, defined by a pair of positions, these being the bottom left and top right, respectively.

colisoesPersonagens

Arguments

:: Personagem
Entity 1

-> Personagem
Entity 2

-> Bool
Logical value of the collision between the two entities

The colisoesPersonagens function should test whether two characters collide, i.e. whether the hitbox of the characters in question intersects in any way. To define it, as well as using the auxiliary function hitPersona mentioned above, we also use:

colPersona: Function that checks if two hitboxes intersect

Properties:

Main function "colisoesPersonagens":

colisoesPersonagens :: Personagem -> Personagem -> Bool

colisoesPersonagens p1 p2 = colPersona (hitPersona p1) (hitPersona p2)

Auxiliary function "colPersona"

colPersona :: Hitbox -> Hitbox -> Bool

colPersona ((xi1, yi1), (xi2, yi2)) ((xe1, yd1), (xe2, yd2)) =

elem True [x >= min xe1 xe2 && x <= max xe1 xe2 && y >= min yd1 yd2 && y <= max yd1 yd2 | (x, y) <- v]

where

v = [(x, y) | x <- [xi1, xi2], y <- [yi1, yi2]]

Examples of use:

>>> colisoesPersonagens (Personagem (1, 0) Jogador (1.5, 1.5) Este (1.5, 1.5) False False 3 0 (False, 0)) (Personagem (1, 0) Jogador (0.5, 0.5) Este (1, 1) False False 3 0 (False, 0))
True

>>> colisoesPersonagens (Personagem (1, 0) Jogador (0.5, 0.5) Este (1, 1) False False 3 0 (False, 0)) (Personagem (1, 0) Jogador (2.5, 2.5) Este (1.5, 1.5) False False 3 0 (False, 0))
False

colPersona

Arguments

:: Hitbox
Character's hitbox

-> Hitbox
Character's hitbox

-> Bool
Logical value of the collision between two entities

Function that checks if two hitboxes intersect