Task2

Description

Module for the realisation of Task2 of LI1 in 2023/24.

Documentation

valida

Arguments

:: Jogo
Game Definition

-> Bool
Logical value of checking all the requested conditions

In this task, we aim to define the valida function, a function that checks whether a given game violates the following restrictions:

The map has a "chao", i.e. a platform that prevents the player or another character from falling off the map.

All the enemies have the ressalta property at True, while the player has it at False.

A player's starting position cannot collide with another character's starting position. Note that the starting positions of enemies can collide with each other.

Minimum number of enemies: 2 (two).

Fantasma enemies have exactly 1 (one) life.

Ladders cannot start/end at trapdoors, and at least one of their ends must be of the Plataforma type.

Trap doors cannot be wider than the player.

There can be no characters or collectibles "inside" platforms or trapdoors, i.e. the block (in the map matrix) corresponding to the position of a character or object must be Vazio.

To be able to define this function, we can use auxiliary functions that check the logical value of each of these properties, such as:

chao function, which checks whether a map has a "floor", i.e. a platform that prevents the player from falling off the map.

ressalto function, which checks that all enemies have the "ressalta" property True and the player False.

choquePs function, which checks that the player's starting position does not coincide with another character's starting position.

minInimigos function, which checks that the minimum number of enemies on the map is 2 (two).

vidaFan function, which checks that all enemies of the Fanstasma type have exactly 1 (one) life.

checkBloc function, which checks that all the ladders on the map don't start or end in trapdoors and that at least one of their ends is of type Platforma.

alcLargo function, which checks if a player is smaller than a map block.

colEntidades function, which checks if a character or collectible is "inside" an empty map block.

In the case of the choquePs function, we use:

the colp function, which removes a character's position.

In the case of the vidaFan function, we use an auxiliary function vidaFanAux, which checks that all ghost-type enemies have exactly 1 (one) life. The main function only checks that all cases are of type True.

In the case of the checkBloc function, we've used auxiliary functions already defined previously, such as findBloc, adapted to detect the Escada type (findBlocE, with the help of findBlocEAux to be able to use it by giving a Mapa as an argument), and created new ones such as:

function extEscada, which receives the matrix that defines the map and the list of ladder block positions on the map (calculated by findblocE), and returns a list of pairs that correspond to the block above and below the ladders, respectively.

function checkBLocAux function, which receives a list of pairs of blocks above and below the stairs and returns the list of logical values for whether the stairs have at least one platform block and no trapdoor blocks.

In the case of the colEntidades function, we have used auxiliary functions already defined previously, such as FindBloc, adapted to detect the Vazio type (findBlocV), and created new ones such as:

function colEntidadesAux, which calculates the list of positions of all entities, simultaneously.

centra function, which calculates the coordinates of the centre of a block.

Properties:

Main function "valida":

valida :: Jogo -> Bool

valida (Jogo m i c p)

= (all (==True) [f1,f2,f3,f4,f5,f6,f7,f8])

where

f1 = chao m

f2 = ressalto i p

f3 = not (choquePs m i)

f4 = minInimigos i

f5 = vidaFan (vidaFanAux i)

f6 = checkBloc m (extEscada m (findBlocEAux m))

f7 = alcLargo p

f8 = colEntidades i p c m

Auxiliary function "chao":

chao :: Mapa -> Bool

chao (Mapa _ _ m) = not (notElem Plataforma (last m))

Auxiliary function "ressalto":

ressalto :: [Personagem] -> Personagem -> Bool

ressalto li (Personagem _ _ _ _ _ _ c2 _ _ _)

= all (== True) (map colc li) && c2 == False

where

colc :: Personagem -> Bool

colp (Personagem _ _ p2 _ _ _ _ _ _ _) = p2

Auxiliary function "choquePs":

choquePs :: Mapa -> [Personagem] -> Bool

choquePs (Mapa (p1,_) _ _) li

= any (== (acC p1)) (map colp li)

where

colp :: Personagem -> Posicao

colp (Personagem _ _ p2 _ _ _ _ _ _ _) = acC p2

Auxiliary function "minInimigos":

minInimigos :: [Personagem] -> Bool

minInimigos i = length i >= 2

Auxiliary function "vidaFan":

vidaFan :: [Bool] -> [Bool]

vidaFan l = all (True==) l

Auxiliary function "vidaFanAux":

vidaFanAux [] = []

vidaFanAux ((Personagem _ t _ _ _ _ _ v _ _):ps)

| t == MacacoMalvado || t == Jogador = True : vidaFanAux ps

| t == Fantasma && v == 1 = True : vidaFanAux ps

| otherwise = False : vidaFanAux ps

Auxiliary function "checkBloc":

checkBloc :: Mapa -> [(Bloco, Bloco)] -> Bool

checkBloc (Mapa _ _ m) l = if findBlocE m == [] then True else if not (null (findBlocE m)) && length l >= 1 then all (==True) (checkBlocAux l) else False

Auxiliary function "checkBlocAux":

checkBlocAux :: [(Bloco,Bloco)] -> [Bool]

checkBlocAux [] = [True]

checkBlocAux ((x,y):r)

| x == Plataforma && (y == Plataforma || y == Vazio || y == Escada) = True : checkBlocAux r

| (x == Plataforma || x == Vazio || x == Escada) && y == Plataforma = True : checkBlocAux r

| otherwise = False : checkBlocAux r

Auxiliary function "extEscada":

extEscada :: Mapa -> [Posicao] -> [(Bloco, Bloco)]

extEscada _ [] = []

extEscada (Mapa i1 i2 m) ((l, c):re)

| l > 0 && l < fromIntegral (length m) - 1 && c >= 0 && c < fromIntegral (length (head m)) =

let lAn = floor (l - 1)

lAt = floor l

lS = floor (l + 1)

cA = floor c

cima = m !! lAn !! cA

baixo = m !! lS !! cA

ext = (cima, baixo)

r = extEscada (Mapa i1 i2 m) re

in if cima /= Alcapao && baixo /= Alcapao && (cima == Plataforma || baixo == Plataforma)

then ext : r

else r

| otherwise = extEscada (Mapa i1 i2 m) re

Auxiliary function "alcLargo":

alcLargo :: Personagem -> Bool

alcLargo (Personagem _ _ _ _ (g,c) _ _ _ _ _) = g <= 1 && c <= 1

Auxiliary function "colEntidades": --

colEntidades :: [Personagem] -> Personagem -> [(Colecionavel, Posicao)] -> Mapa -> Bool

colEntidades ps p cs (Mapa i1 i2 m) = all (`elem` (centra ((findBlocV m) ++ (findBlocE m)))) (colEntidadesAux ps p cs)

Auxiliary function "colEntidadesAux":

colEntidadesAux :: [Personagem] -> Personagem -> [(Colecionavel, Posicao)] -> [Posicao]

colEntidadesAux ps (Personagem _ _ (x2, y2) _ _ _ _ _ _ _) cs =

map (acC . (\(Personagem _ _ (x1, y1) _ _ _ _ _ _ _)-> (x1, y1))) ps ++ [acC (x2, y2)] ++ map snd cs

Auxiliary function "centra":

centra :: [Posicao] -> [Posicao]

centra [] = []

centra m = map (\(l, ls) -> (l + 0.5, ls + 0.5)) m

Examples of use:

>>> valida (Jogo Mapa ((1.5,1.5),Este) (1.5,1.5) [[Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma]] [(Personagem (1, 0) Fantasma (2.5, 1.5) Este (1, 1) False True 1 0 (False, 0)),(Personagem (1, 0) Fantasma (2.5, 1.5) Este (1, 1) False True 1 0 (False, 0))] [(Moeda, (1.5,1.5))] Personagem (1, 0) Jogador (0.5, 0.5) Este (1, 1) False False 3 0 (False, 0))
True

>>> valida (Jogo Mapa ((1.5,1.5),Este) (1.5,1.5) [[Vazio, Alcapao, Vazio], [Vazio, Escada, Vazio], [Vazio, Vazio, Plataforma]] [(Personagem (1, 0) Fantasma (1.5, 1.5) Este (1, 1) False False 2 0 (False, 0))] [(Martelo,(2.5,2.5))] Personagem (1, 0) Jogador (2.5, 2.5) Este (2, 1) False True 3 0 (False, 0))
False

chao

Arguments

:: Mapa
Game map

-> Bool
Logical value of the Map floor being made entirely of Plataforma blocks

Function, which checks whether a map has a "floor", i.e. a platform that prevents the player from falling off the map.

ressalto

Arguments

:: [Personagem]
Enemies List

-> Personagem
Player character

-> Bool
Logical value of the "ressalta" property to be True on enemies and False on the Player

Function, which checks that all enemies have the "ressalta" property True and the player False.

choquePs

Arguments

:: Mapa
Game map

-> [Personagem]
Enemies list

-> Bool
Logical value of the enemies' spawning place being different from the Player's spawn

Function that checks that the player's starting position does not coincide with the starting position of another character.

minInimigos

Arguments

:: [Personagem]
Enemies list

-> Bool
Logical value of the minimum number of enemies on the map being more or less than two

Function, which checks that the minimum number of enemies on the map is 2 (two).

vidaFan

Arguments

:: [Bool]
List of logical values calculated in "vidaFan"

-> Bool
Logic value of all list

Function that checks if all the cases studied in the vidaFanAux function are true.

vidaFanAux

Arguments

:: [Personagem]
Enemies list on map

-> [Bool]
List of logical values if the life of the Fantasma type is equal to 1

Function that checks if all enemies of type Fantasma have exactly 1 (one) life.

checkBloc

Arguments

:: Mapa
Game map

-> [(Bloco, Bloco)]
List of pairs that correspond to the block above and below the ladder, respectively.

-> Bool

Function that checks that all the stairs on the map do not start or end in trapdoors, and that at least one of their ends is of type Plataforma.

checkBlocAux

Arguments

:: [(Bloco, Bloco)]
List of pairs that correspond to the block above and below the ladder, respectively.

-> [Bool]
List of logical values.

Function that receives a list of pairs of blocks above and below the stairs and returns the list of logical values of whether the stairs have at least one platform block and no trapdoor blocks.

extEscada

Arguments

:: Mapa
Game map

-> [Posicao]
List of ladder block positions

-> [(Bloco, Bloco)]
List of pairs that correspond to the block above and below the ladder, respectively.

Function that receives the matrix that defines the map and the list of ladder block positions on the map (calculated by findblocE), and returns a list of pairs that correspond to the block above and below the ladders, respectively.

findBlocEAux

Arguments

:: Mapa
Game map

-> [Posicao]
List of coordinates in the top left corner of the block

findBlocE

Arguments

:: [[Bloco]]
Map definition matrix

-> [Posicao]
List of coordinates in the top left corner of the block

alcLargo

Arguments

:: Personagem
Player character

-> Bool
Logical value of the Player Character being smaller than a Trapdoor block on the map

Function that checks if a player is smaller than a Trapdoor block on the map.

colEntidades

Arguments

:: [Personagem]
Enemies List

-> Personagem
Player character

-> [(Colecionavel, Posicao)]
Collectibles list

-> Mapa
Map definition matrix

-> Bool
Logical value of a character or collectible being "inside" an empty map block

Function that checks if any character or collectible is "inside" an Empty block on the map.

findBlocV

Arguments

:: [[Bloco]]
Map definition matrix

-> [Posicao]
List of coordinates in the top left corner of the block

centra

Arguments

:: [Posicao]
List of coordinates in the top left corner of the block

-> [Posicao]
List of coordinates in the center of the block

Function that calculates the coordinates of the centre points of a block.

colEntidadesAux

Arguments

:: [Personagem]
Enemies list

-> Personagem
Player character

-> [(Colecionavel, Posicao)]
Collectibles list

-> [Posicao]
List of positions of all entities on the map

Function that calculates the list of positions of all entities simultaneously.