Task3

Description

Module for the realisation of Task3 of LI1 in 2023/24.

Documentation

movimenta

Arguments

:: Semente
Game seed

-> Tempo
Game time

-> Jogo
Game

-> Jogo
Updated game

The movimetna function must animate all the characters, i.e. calculate their new positions and their consequences. To do this, the function has to test a non-exaustive list of events and conditions. To do this, we use auxiliary functions whose purpose is to check conditions and make changes to the game, such as:

perderVida: function that receives the list of enemies on a map and returns this same list with the lives removed from the supposed enemies.

gamaIni: function that filters out enemies with 1 life and suppresses those with 0 lives.

menosAl: function that updates the game map matrix when a trapdoor is stepped on.

fallVoid: function that changes the character's speed if they are not on a platform.

danoDuro: function that removes a life from the player if they come into contact with an enemy.

recColec: function that updates the game when a collectible is acquired, and depending on the collectible makes changes to either the player's points or the player's character properties.

noPassPar: function that slows the player down horizontally and/or vertically, depending on their collision with the edges of the map or with platform blocks.

Properties:

Main function "movimenta":

movimenta :: Semente -> Tempo -> Jogo -> Jogo

movimenta s t j@(Jogo m@(Mapa i1 i2 i3) i c p) =

Jogo mr (movInimigo t (randInimigos m s (gamaIni (perderVida i p)))) co (movJogador t (danoDuro i (noPassPar m (fallVoid m jo))))

where

mr = (Mapa i1 i2 (menosAl m p))

ri = movInimigo t (randInimigos m s i)

(co,jo) = recColec j

Auxiliary function "perderVida":

perderVida :: [Personagem] -> Personagem -> [Personagem]

perderVida [] _ = []

perderVida ((Personagem i1 i2 i3 i4 i5 i6 i7 v i9 (d,id2)):ls) p

| acH (hitPersona (Personagem i1 i2 i3 i4 i5 i6 i7 v i9 (d,id2))) == hitDano p && (danAtivo p == True) = ((Personagem i1 i2 i3 i4 i5 i6 i7 (v - 1) i9 (d,id2))) : perderVida ls p

| otherwise = (Personagem i1 i2 i3 i4 i5 i6 i7 v i9 (d,id2)) : perderVida ls p

Auxiliary function "hitDano":

hitDano :: Personagem -> Hitbox

hitDano (Personagem i1 i2 (x, y) o (g, c) i3 i4 i5 i6 i7)

| o == Este   = (somPar (fst (acH (hitPersona p))) (1,0), somPar (snd (acH (hitPersona p))) (1,0))

| o == Oeste  = (somPar (fst (acH (hitPersona p))) (-1,0), somPar (snd (acH (hitPersona p))) (-1,0))

| otherwise   = acH (hitPersona (Personagem i1 i2 (x, y) o (g, c) i3 i4 i5 i6 i7))

where

p = (Personagem i1 i2 (x, y) o (g, c) i3 i4 i5 i6 i7)

Auxiliary function "danAtivo":

danAtivo :: Personagem -> Bool

danAtivo (Personagem _ _ _ _ _ _ _ _ _ (d,_))

= if d == True

then True

else False

Auxiliary function "gamaIni":

gamaIni :: [Personagem] -> [Personagem]

gamaIni i = filter (\(Personagem _ _ _ _ _ _ _ v _ _) -> v > 0) i

Auxiliary function "menosAl":

menosAl :: Mapa -> Personagem -> [[Bloco]]

menosAl mp@(Mapa _ _ m) p@(Personagem _ e (x, y) _ _ _ _ _ _ _)

| e == Jogador && elem (fst (hitPersona (acP p))) (findBlocA m) = substituiMatriz m (yf,xf) Vazio

| otherwise = m

where (xf,yf) = fst (hitPersona (acP p))

Auxiliary function "substituiMatriz":

substituiMatriz :: [[Bloco]] -> Posicao -> Bloco -> [[Bloco]]

substituiMatriz [] _ _ = []

substituiMatriz (l:t) (0,c) v = (substitui l c v) : t

substituiMatriz (l:t) (linha, coluna) v = l : substituiMatriz t (linha - 1, coluna) v

Auxiliary function "substitui":

substitui :: [Bloco] -> Double -> Bloco -> [Bloco]

substitui [] _ _ = []

substitui (x:t) 0 y = y:t

substitui (x:t) p y = x : substitui t (p-1) y

Auxiliary function "findBlocA":

findBlocA :: [[Bloco]] -> [Posicao]

findBlocA [] = []

findBlocA m =

concatMap (\(l, bs) -> [(fromIntegral c, fromIntegral l) | (b, c) <- zip bs [0..], b == Alcapao]) $ zip [0..] m

Auxiliary function "fallVoid":

fallVoid :: Mapa -> Personagem -> Personagem

fallVoid (Mapa _ _ b) p@(Personagem (vx,vy) _ (x, y) _ _ i6 _ _ _ _)

| fst (hitPersona (acP p)) `elem` findBlocP b = p

| i6 = p

| otherwise = p { velocidade = (vx, vy + 10) }

Auxiliary function "danoDuro":

danoDuro :: [Personagem] -> Personagem -> Personagem

danoDuro [] p = p

danoDuro l p@(Personagem _ _ _ _ _ _ _ x _ _) | any (==True) (colisoesDano l p) = p { vida = (x - 1) }

| otherwise = p

Auxiliary function "colisoesDano":

colisoesDano :: [Personagem] -> Personagem -> [Bool]

colisoesDano [] p = [False]

colisoesDano (h:ht) p = colisoesPersonagens h p : colisoesDano ht p

Auxiliary function "recColec":

recColec :: Jogo -> Jogo

recColec jogo@(Jogo m@(Mapa m1 m2 m3) i [] p) = ([],p)

recColec (Jogo m i (colec@(c,pov):r) p@(Personagem _ _ pos _ _ _ _ _ moeda dano))

| colPersona (acH (hitColec pov)) (acH (hitPersona p)) && c == Martelo = (r, p { aplicaDano = (True,10) } )

| colPersona (acH (hitColec pov)) (acH (hitPersona p)) && c == Moeda   = (r, p { pontos = (moeda + 1)} )

| otherwise = insColec colec (recColec (Jogo m i r p))

Auxiliary function "insColec":

insColec :: (Colecionavel, Posicao) -> ([(Colecionavel, Posicao)],Personagem) -> ([(Colecionavel, Posicao)],Personagem)

insColec c (r,p) = (c:r, p)

Auxiliary function "hitColec":

hitColec :: Posicao -> Hitbox

hitColec (x,y)

= ((xe1,yd1),(xe2,yd2))

where

xe1 = x - (1 / 2)

yd1 = y + (1 / 2)

xe2 = x + (1 / 2)

yd2 = y - (1 / 2)

Auxiliary function "noPassPar":

noPassPar :: Mapa -> Personagem -> Personagem

noPassPar mp@(Mapa _ _ m) p@(Personagem (vx, vy) _ (x, y) _ (g, c) _ _ _ _ _)

|  t1 && t2 = p {velocidade = (0,0)}

|  t1 = p {velocidade = (0,vy)}

|  t2 = p {velocidade = (vx,0)}

|  otherwise = p

where

t1 = colPar m p

t2 = any (`elem` [(x2,y2)]) (findBlocP m)

where

(x2,y2) = fst (hitPersona p)

Auxiliary function "movJogador":

movJogador :: Personagem -> Personagem

movJogador t (Personagem (vx,vy) i2 (x, y) i4 i5 i6 i7 i8 i9 i10) =

(Personagem (0,0) i2 (x + (vx * t), y + (vy * t)) i4 i5 i6 i7 i8 i9 i10)

Auxiliary function "movInimigo":

movInimigo :: [Personagem] -> [Personagem]

movInimigo t i = map (movJogador t) i

Auxiliary function "randInimigos":

randInimigos m s [] = []

randInimigos m s [] = []

randInimigos m s (i:is) = randInimigosAux m i (last l) : randInimigos m s is

where

l = geraAleatorios s (length (i:is))

Auxiliary function "randInimigosAux"

randInimigosAux :: Mapa -> Personagem -> Int -> Personagem

randInimigosAux m@(Mapa m1 m2 m3) i@(Personagem (v1,v2) i2 i3 i4 i5 i6 i7 i8 i9 i10) n =

case n `mod` 4 of

0 -> aD m3 i

1 -> aE m3 i

2 -> aS m3 i

3 -> aC m3 i

_ -> i

where

aD :: [[Bloco]] -> Personagem -> Personagem

aD m3 i = if movDireita m3 i then (Personagem (v1-1,v2) i2 i3 Oeste i5 i6 i7 i8 i9 i10) else (Personagem (v1+1,v2) i2 i3 Este i5 i6 i7 i8 i9 i10)

aE :: [[Bloco]] -> Personagem -> Personagem

aE m3 i = if movEsquerda m3 i then (Personagem (v1+1,v2) i2 i3 Este i5 i6 i7 i8 i9 i10) else (Personagem (v1-1,v2) i2 i3 Oeste i5 i6 i7 i8 i9 i10)

aS :: [[Bloco]] -> Personagem -> Personagem

aS m3 i = if subEscada m3 i then (Personagem (v1,v2-1) i2 i3 i4 i5 i6 i7 i8 i9 i10) else i

aC :: [[Bloco]] -> Personagem -> Personagem

aC m3 i = if subEscada m3 i then (Personagem (v1,v2+1) i2 i3 i4 i5 i6 i7 i8 i9 i10) else i

Examples of use:

>>> movimenta (fromIntegral 1) 2 (Jogo (Mapa ((2.5, 1.5), Este) (5, 5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Alcapao, Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]]) ([(Personagem (1, 0) Fantasma (2.5, 1.5) Oeste (1, 1) False True 1 0 (False, 0))]) ([(Moeda, (1.5,1.5)), (Martelo,(4.5,4.5))]) (Personagem (0, 0) Jogador (1.5, 1.5) Este (1, 1) False True 3 0 (True, 5)))
Jogo Mapa ((2.5, 1.5), Este) (5, 5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]] [] [(Martelo,(4.5,4.5))] Personagem (0, 10) Jogador (1.5, 1.5) Este (1, 1) False True 2 1 (True, 5)

>>> movimenta (fromIntegral 1) 2 (Jogo (Mapa ((2.5, 1.5), Este) (5, 5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Plataforma, Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]]) ([(Personagem (1, 0) Fantasma (3.5, 3.5) Oeste (1, 1) False True 1 0 (False, 0))]) ([(Martelo,(4.5,4.5))]) (Personagem (0, 0) Jogador (1.5, 1.5) Este (1, 1) False True 3 0 (False, 0)))
Jogo Mapa ((2.5, 1.5), Este) (5, 5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Plataforma, Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]] [(Personagem (1, 0) Fantasma (3.5, 3.5) Oeste (1, 1) False True 1 0 (False, 0))] [(Martelo,(4.5,4.5))] Personagem (0, 0) Jogador (1.5, 1.5) Este (1, 1) False True 3 0 (False, 0)

Function that must animate all the characters, i.e. calculate their new positions and their consequences.

movJogador

Arguments

:: Tempo
Game time

-> Personagem
Player character

-> Personagem
Updated player character

Function that changes the position of the player's character according to speed

movInimigo

Arguments

:: Tempo
Game time

-> [Personagem]
Enemies list

-> [Personagem]
Updated enemies list

Function that changes the position of enemies according to their speed

randInimigos

Arguments

:: Mapa
Game map

-> Semente
Game seed

-> [Personagem]
Enemies list

-> [Personagem]
Updated enemies list

Function that generates random movement to a list of enemies

randInimigosAux

Arguments

:: Mapa
Game map

-> Personagem
Game character

-> Int
Randomly generated 18-digit number

-> Personagem
Updated character

Function that assists "randEnemies"

perderVida

Arguments

:: [Personagem]
Enemies list

-> Personagem
Player character

-> [Personagem]
Updated enemies list

Function that receives the list of enemies on a map and returns this same list with the lives taken from the supposed enemies.

hitDano

Arguments

:: Personagem
Player character

-> Hitbox
Player damage hitbox

Function that calculates the damage hitbox of the Player's character.

somPar :: Posicao -> Posicao -> Posicao

danAtivo

Arguments

:: Personagem
Player character

-> Bool
Logical value of the "aplicaDano" property being active

function that checks if the "aplicaDano" property on the Player character is _False_

gamaIni

Arguments

:: [Personagem]
Enemies list

-> [Personagem]
List of enemies with 1 (one) life

Function that filters out enemies with 1 life and suppresses those with 0 lives.

fallVoid

Arguments

:: Mapa
Game map

-> Personagem
Player character

-> Personagem
Player character with updated speed.

Function that changes the character's speed if he is not on a platform.

danoDuro

Arguments

:: [Personagem]
Enemies list

-> Personagem
Player character

-> Personagem
Player character with updated life.

Function that removes a life from the player if they come into contact with an enemy.

colisoesDano

Arguments

:: [Personagem]
Enemies list

-> Personagem
Player character

-> [Bool]
List of the truth values of enemies in contact with the player

Function that checks if any enemies are colliding with the player.

recColec

Arguments

:: Jogo
Game

-> ([(Colecionavel, Posicao)], Personagem)
Updated collectibles

Function that updates the game when a collectible is acquired, and depending on the collectible makes changes to either the player's points or the player's character properties.

insColec

Arguments

:: (Colecionavel, Posicao)
Collectible and its previously processed position

-> ([(Colecionavel, Posicao)], Personagem)
Remaining collectibles and character

-> ([(Colecionavel, Posicao)], Personagem)
Complete (updated) list of collectibles and characters

hitColec

Arguments

:: Posicao
Position on the map of a collectible

-> Hitbox
Collectible hitbox

Function that calculates the hitbox of a collectible.

menosAl

Arguments

:: Mapa
Game map

-> Personagem
Player character

-> [[Bloco]]
Map matrix with updated trapdoors

Function that updates the game map matrix when a trapdoor is stepped on.

substituiMatriz

Arguments

:: [[Bloco]]
Map Matrix

-> Posicao
Position of a block on the map

-> Bloco
Block

-> [[Bloco]]
Matrix with updated block

Function that replaces a block in an array with another block.

substitui

Arguments

:: [Bloco]
List of blocks

-> Double
Position on the list

-> Bloco
Block

-> [Bloco]
Updated list

Function that replaces a block in a list of blocks with another block.

findBlocA

Arguments

:: [[Bloco]]
Map definition matrix

-> [Posicao]
List of coordinates for the top left corner of the trapdoor block

Function that calculates the list of positions of all the trapdoor blocks in a map matrix.

noPassPar

Arguments

:: Mapa
Game map

-> Personagem
Player character

-> Personagem
Player character with updated speed.

Function that slows the player down horizontally and/or vertically, depending on their collision with the edges of the map or platform blocks.