Task4

Description

Module for the realisation of Task4 of LI1 in 2023/24.

Documentation

atualiza

Arguments

:: [Maybe Acao]
Actions to apply to enemies

-> Maybe Acao
Action to be applied to the player

-> Jogo
Game

-> Jogo
Updated Game

The atualiza function must validate the new directions and speeds of the characters (enemies and player) according to the actions given. To define it, we need to use the following auxiliary functions:

atualizaIni: Function that updates the enemies according to the list of actions given

atualizaPer: Function that updates the player character according to the given action

To define these two auxiliary functions, we use other functions that will be called by them, such as:

atualizaAcao: Function that, given an action, updates a character's data.

subEscada: Function that checks if a character is in the position of a Ladder type block.

movDireita: Function that checks if a character cannot move to the right

movEsquerda: Function that checks if a character cannot move to the left

parOp: Function that checks if a character can be stationary

Properties:

Main function "atualiza":

atualiza :: [Maybe Acao] -> Maybe Acao -> Jogo -> Jogo

atualiza ap aj j@(Jogo m i c p) =

let

ni = atualizaIni m ap i

nj = atualizaPer m aj p

in

(Jogo m ni c nj)

Auxiliary function "atualizaIni":

atualizaIni :: Mapa -> [Maybe Acao] -> [Personagem] -> [Personagem]

atualizaIni _ [] i = i

atualizaIni m (ac:acs) i =

case ac of

Just ac  -> atualizaIni m acs (map (atualizaAcao m ac) i)

Nothing  ->

Auxiliary function "atualizaPer":

atualizaPer :: Mapa -> Maybe Acao -> Personagem -> Personagem

atualizaPer m mac p =

case mac of

Just ac -> atualizaAcao m ac p

Nothing -> p

Auxiliary function "atualizaAcao":

atualizaAcao :: Mapa -> Acao -> Personagem -> Personagem

atualizaAcao m@(Mapa m1 m2 m3) ac p@(Personagem (v1,v2) i2 i3 i4 i5 i6 i7 i8 i9 i10) =

case ac of

Subir         -> if subEscada m3 pm then Personagem (v1,v2-1) i2 i3 i4 i5 i6 i7 i8 i9 i10 else p

Descer        -> if subEscada m3 pm then Personagem (v1,v2+1) i2 i3 i4 i5 i6 i7 i8 i9 i10 else p

AndarDireita  -> if movDireita m3 pm then Personagem (v1-1,v2) i2 i3 Oeste i5 i6 i7 i8 i9 i10 else Personagem (v1+1,v2) i2 i3 Este i5 i6 i7 i8 i9 i10

AndarEsquerda -> if movEsquerda m3 pm then Personagem (v1+1,v2) i2 i3 Este i5 i6 i7 i8 i9 i10 else Personagem (v1-1,v2) i2 i3 Oeste i5 i6 i7 i8 i9 i10

Saltar        -> if not i6 then Personagem (v1,v2-1) i2 i3 i4 i5 i6 i7 i8 i9 i10 else p

Parar         -> if parOp m3 pm then Personagem (0,0) i2 i3 i4 i5 i6 i7 i8 i9 i10 else p

where

pm = acP p

Auxiliary function "subEscada":

subEscada :: [[Bloco]] -> Personagem -> Bool

subEscada b p@(Personagem i1 i2 i3 i4 i5 s i7 i8 i9 i10)

| s = True

| any (== hitPersona p) (hitBlocos (findBlocE b)) = True

| otherwise = False

Auxiliary function "movDireita":

movDireita :: [[Bloco]] -> Personagem -> Bool

movDireita b p@(Personagem i1 i2 (x, y) d i5 i6 r i8 i9 i10)

| r && (`elem` findBlocP b) (snd (hitPersona p)) = True

| r && (`elem` findBlocV b) (fst (fst (hitPersona p)) + 1, snd (fst (hitPersona p))) = True

| otherwise = False

Auxiliary function "movEsquerda":

movEsquerda :: [[Bloco]] -> Personagem -> Bool

movEsquerda b p@(Personagem i1 i2 (x, y) d i5 i6 r i8 i9 i10)

| r && (`elem` findBlocP b) (fst (snd (hitPersona p)) - 2, snd (snd (hitPersona p))) = True

| r && (`elem` findBlocV b) (fst (fst (hitPersona p)) - 1, snd (fst (hitPersona p))) = True

| otherwise = False

Auxiliary function "parOp":

parOp :: [[Bloco]] -> Personagem -> Bool

parOp b p@(Personagem i1 i2 (x, y) d i5 i6 r i8 i9 i10)

| (`elem` findBlocP b) (fst (hitPersona p)) = True

| otherwise = False

Examples of use:

>>> atualiza ([Just AndarDireita, Just AndarDireita]) ((Just Subir)) (Jogo (Mapa ((1.5, 1.5), Este) (5, 5) [[Escada, Vazio, Vazio], [Vazio, Vazio, Plataforma], [Plataforma, Plataforma, Plataforma]]) ([(Personagem (0, 0) Fantasma (1.5, 1.5) Este (1, 1) False True 1 0 (False, 0)),(Personagem (0, 0) Fantasma (0.5, 1.5) Oeste (1, 1) False False 1 0 (False, 0))]) ([(Moeda,(1.5,1.5)),(Martelo,(1.5,1.5))]) (Personagem (5, 0) Jogador (0.5, 0.5) Oeste (1, 1) True True 2 0 (True, 5)))
Jogo (Mapa ((1.5, 1.5), Este) (5, 5) [[Escada, Vazio, Vazio], [Vazio, Vazio, Plataforma], [Plataforma, Plataforma, Plataforma]]) ([(Personagem (-2, 0) Fantasma (1.5, 1.5) Oeste (1, 1) False True 1 0 (False, 0)),(Personagem (2, 0) Fantasma (0.5, 1.5) Este (1, 1) False False 1 0 (False, 0))]) ([(Moeda,(1.5,1.5)),(Martelo,(1.5,1.5))]) (Personagem (5, -1) Jogador (0.5, 0.5) Oeste (1, 1) True True 2 0 (True, 5))

>>> atualiza ([Nothing]) (Nothing) Jogo (Mapa ((2.5, 1.5), Este) (5, 5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Alcapao, Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]]) ([(Personagem (1, 0) Fantasma (3.0, 1.5) Oeste (1, 1) False True 1 0 (False, 0)),(Personagem (1, 0) Fantasma (0.5, 1.5) Oeste (1, 1) False True 1 0 (False, 0)),(Personagem (1, 0) Fantasma (2.5, 2.5) Oeste (1, 1) False True 0 0 (False, 0))]) ([(Martelo,(4.5,4.5))]) (Personagem (0, 0) Jogador (0.5, 1.5) Oeste (1, 1) False True 2 0 (True, 5))
Jogo (Mapa ((2.5, 1.5), Este) (5, 5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Alcapao, Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]]) ([(Personagem (1, 0) Fantasma (3.0, 1.5) Oeste (1, 1) False True 1 0 (False, 0)),(Personagem (1, 0) Fantasma (0.5, 1.5) Oeste (1, 1) False True 1 0 (False, 0)),(Personagem (1, 0) Fantasma (2.5, 2.5) Oeste (1, 1) False True 0 0 (False, 0))]) ([(Martelo,(4.5,4.5))]) (Personagem (0, 0) Jogador (0.5, 1.5) Oeste (1, 1) False True 2 0 (True, 5))

Function that validates the new directions and speeds of the characters (enemies and player) according to the actions taken.

atualizaIni

Arguments

:: Mapa
Game map

-> [Maybe Acao]
List of actions

-> [Personagem]
Enemies List

-> [Personagem]
Updated enemies list

Function that updates enemies according to the list of actions given.

atualizaPer

Arguments

:: Mapa
Game map

-> Maybe Acao
Action to be applied

-> Personagem
Player character

-> Personagem
Player character updated after action

Function that updates the player's character according to the action taken.

atualizaAcao

Arguments

:: Mapa
Game map

-> Acao
Action to be applied

-> Personagem
Character

-> Personagem
Updated character

Function that, given an action, updates a character's data.

subEscada

Arguments

:: [[Bloco]]
Map definition matrix

-> Personagem
Character

-> Bool
Truth value of the character being in the position of a Ladder block

Function that checks whether a character is in the position of a Ladder type block.

movDireita

Arguments

:: [[Bloco]]
Map definition matrix

-> Personagem
Character

-> Bool
True value of the character not being able to move to the right

Function that checks if a character cannot move to the right.

movEsquerda

Arguments

:: [[Bloco]]
Map definition matrix

-> Personagem
Character

-> Bool
rue value of the character not being able to move to the left

Function that checks if a character cannot move to the left.

parOp

Arguments

:: [[Bloco]]
Map definition matrix

-> Personagem
Character

-> Bool
Truth value of the character being able to stand still

Function that checks whether a character can be stationary.