Task5

Description

Module for the realisation of Task5 of LI1 in 2023/24.

Documentation

data Menu

In this task we carried out the graphics component of the game, which allowed us to transpose all the previous tasks into a playable executable, based on the arcade game "Donkey Kong". To do this, we defined new dates and types, and carried out all the necessary processes to load the images that will make up the game graphics, from the menus to the collectibles, enemies, etc.

The various menus in the game

Constructors

EmJogo1

MenuInicial

MCreditos

SelecMapa

PauseM

WinM

LoseM

Sure

MenuPkp

PauseP

WinP

LoseP

EmJogo2

data Opcoes

The various options in the game

Constructors

Limpo

Jogar

Sair

Creditos

Retornar

Retomar

NextMap

SelMap

Mapa1

Mapa2

Mapa3

Plus

data Temas

Game themes/modes

Constructors

PlusK

Classico

data Icon

The various auxiliary icons used during the game

Constructors

TimerM

MoedaP

VidaI

Estrelinha

type ImagensBlocos = [(Temas, [(Bloco, Picture)])]

The type used to describe the images of the map blocks

type ImagensPersonagens = [(Temas, [(Entidade, Pictures)])]

The type used to describe the images of the characters in the game

type ImagensMenus = [(Menu, Pictures)]

The type used to describe the images in the game's menus

type ImagensColecionaveis = [(Temas, [(Colecionavel, Picture)])]

The type used to describe the images of the game's collectibles

type ImagensIcones = [(Icon, Pictures)]

The type used to describe the images of the game icons

type Pictures = [Picture]

The type used to summarise a list of Pictures

data Imagens

The date responsible for gathering all types of images into a compact block

Constructors

Imagens

Fields

blocos :: ImagensBlocos

Describe the images of the map blocks

personagens :: ImagensPersonagens

Describe the images of the characters in the game

imenus :: ImagensMenus

Describe the images of the game menus

icolec :: ImagensColecionaveis

Describe the images of the game collectibles

iicons :: ImagensIcones

Describe the images of the game icons

data Escolha

The date responsible for choosing the map to play in the game

Constructors

PrMapa

SnMapa

TrMapa

Nenhum

data PrimateKong

The date responsible for defining the game itself, with all its parts

Constructors

PrimateKong

Fields

jogo :: Jogo

The map and current entities in the game

menu :: Menu

Game menus

opcao :: Opcoes

Game options

imagens :: Imagens

Images / sprites used in the game

tema :: Temas

Different game themes

semente :: Semente

Seed used for in-game randomisation

tempo :: Tempo

Time running on each map

tecPress :: Tecpress

Used to continue a movement if the key is pressed.

paridade :: Paridade

Used for animations with two frames, analysing time parity

escolham :: Escolha

Chosen map game

gerarSeed

Arguments

:: IO Int
Random value for the game seed

Function used to generate the random number that will form the game seed

gIBlocos

Arguments

:: IO ImagensBlocos
Images of the game blocks

Function used to load images of map blocks

gIPersonagens

Arguments

:: IO ImagensPersonagens
Images of the game's characters

Function used to load images of game entities and their variations

gIMenu

Arguments

:: IO ImagensMenus
Images of the game menus

Function used to load Menu images

gIColecionavel

Arguments

:: IO ImagensColecionaveis
Images of the game's collectibles

Function used to load images of the game's collectibles

gIIcones

Arguments

:: IO ImagensIcones
Images of the game icons

Function used to load the icons used during the game

estadoInicial

Arguments

:: IO PrimateKong
Main game data

Function that returns the initial state of the game, i.e. when you open it

window

Arguments

:: Display
Game Window

Function responsible for dictating the size of the game window as well as its name

backgroundC

Arguments

:: Color
Background color

Function that sets the colour of the game background

fr

Arguments

:: Int
Value of frames per second in the game

Function responsible for defining the frames per second of the game

desenhaEstado

Arguments

:: PrimateKong
Main game data

-> Picture

Function responsible for designing the different game states

desenhaIcon

Arguments

:: ImagensIcones
Icon images

-> Personagem
Player character

-> Tempo
Time in stage

-> Pictures
Images with icons

Function that draws some icons in the game under specific conditions

desenhaMenus

Arguments

:: Menu
Current game menu

-> Opcoes
Current game option

-> ImagensMenus
Menu images

-> Personagem
Player character

-> Tempo
Time in stage

-> ImagensIcones
Icon images

-> Picture
Image composed of the momentarily selected menu

Function responsible for designing all the game's menus, depending on your options

desenhaMapa

Arguments

:: Float
Value in x where you start drawing the map

-> Float
Value in y where you start drawing the map

-> [[Bloco]]
List of map block lists

-> Temas
General themes defined

-> ImagensBlocos
Images of the map blocks

-> Pictures
Map images

Function responsible for drawing the map in general

desenhaLinha

Arguments

:: Float
Value in x where the map line starts to be drawn

-> Float
Value in y where the map line starts to be drawn

-> [Bloco]
List of map blocks

-> Temas
General themes defined

-> ImagensBlocos
Images of the map blocks

-> Pictures
Images of the map lines

Function responsible for drawing the lines on the map

desenhaBloco

Arguments

:: Float
Value in x where the map block begins to be drawn

-> Float
Value in y where the map block begins to be drawn

-> Bloco
Map block in question

-> Temas
General themes defined

-> ImagensBlocos
Images of the map blocks

-> Pictures
Images of the designed blocks

Function responsible for drawing the map blocks

desenhaEntidadeHelp1

Arguments

:: [Personagem]
Enemies list

-> Personagem
Player character

-> ImagensPersonagens
Images of the game's characters

-> Tecpress
Pressing a key

-> Paridade
Stage time parity value

-> Temas
Different game themes

-> Pictures
Images of all the entities in the game

Function responsible for helping to design all the entities in the game

desenhaEntidadeHelp2

Arguments

:: [Personagem]
List of all entities

-> ImagensPersonagens
Images of the game's characters

-> Tecpress
Pressing a key

-> Paridade
Stage time parity value

-> Temas
Different game themes

-> Pictures
Images of all the entities in the game

Function responsible for helping to design all the entities in the game

desenhaEntidade

Arguments

:: Personagem
Character from the game in question

-> ImagensPersonagens
Images of the game's characters

-> Tecpress
Pressing a key

-> Paridade
Stage time parity value

-> Temas
Different game themes

-> Picture
Image of the designed entity

Function responsible for designing one game entity at a time

desenhaColecionavelHelp

Arguments

:: [(Colecionavel, Posicao)]
List of collectibles and their positions in the game

-> ImagensColecionaveis
Images of in-game collectibles

-> Menu
Current Menu

-> Pictures
Images of the game's collectibles

Function responsible for helping design the game's collectibles

desenhaColecionavel

Arguments

:: (Colecionavel, Posicao)
Collectible in question and its position

-> ImagensColecionaveis
Images of in-game collectibles

-> Menu
Current Menu

-> Picture
Image with the designed collectible

Function responsible for designing the game's collectibles

reageEvento

Arguments

:: Event
Key pressed

-> PrimateKong
Compact game data

-> PrimateKong
Compact game data updated after player interaction

Function responsible for linking the player to the game, allowing it to be controlled via keyboard keys

reageTempo

Arguments

:: Float
Time since the game opened

-> PrimateKong
Compact game data

-> PrimateKong
Game data updated according to time

Function responsible for updating the status as a function of time

tempo1 :: Tempo

Time used to restart the counter at each stage

prf :: Tecpress

Used to reset key presses

jogador1 :: Personagem

Definition of the first player

inimigos1 :: [Personagem]

Definition of the first enemies list

colecionaveis1 :: [(Colecionavel, (Double, Double))]

Definition of the first list of collectibles

mapa1 :: Mapa

Definition of the first map

jogo1 :: Jogo

Definition of the first stage

mapa2 :: Mapa

Definition of the second map

jogador2 :: Personagem

Definition of the second player

inimigos2 :: [Personagem]

Definition of the second enemies list

colecionaveis2 :: [(Colecionavel, (Double, Double))]

Definition of the second list of collectibles

jogo2 :: Jogo

Definition of the second stage

mapa3 :: Mapa

Definition of the third map

jogador3 :: Personagem

Definition of the third player

inimigos3 :: [Personagem]

Definition of the third enemies list

colecionaveis3 :: [(Colecionavel, (Double, Double))]

Definition of the third list of collectibles

jogo3 :: Jogo

Definition of the third stage

mapa1b

Arguments

:: [[Bloco]]
Map matrix 1

Definition of the first map matrix

mapa2b

Arguments

:: [[Bloco]]
Map matrix 2

Definition of the second map matrix

mapa3b

Arguments

:: [[Bloco]]
Map matrix 3

Definition of the third map matrix